Fixed extending the paint bounds for non-opaque windows
For windows with alpha < 1.0 or non-opaque windows, need to call ID2D1DeviceContext::Clear to reset the clip region to transparent black before painting. ID2D1DeviceContext::Clear only respects axis-aligned clip regions, so if the clip region is more complex the clip region needs to be extended to the bounds of the deferredAreas rectangle list.